+++ src/Resources/contao/languages/de/tl_content.php
@@ -5,3 +5,4 @@ $GLOBALS['TL_LANG']['tl_content']['kaderliste_legend'] = 'Kaderliste';
 $GLOBALS['TL_LANG']['tl_content']['kaderliste_id'] = array('Liste', 'Bitte wählen Sie eine Liste aus');
 $GLOBALS['TL_LANG']['tl_content']['kaderliste_invisibleDWZ'] = array('DWZ-Spalte ausblenden', 'DWZ-Spalte im Frontend ausblenden');
 $GLOBALS['TL_LANG']['tl_content']['kaderliste_invisibleElo'] = array('Elo-Spalte ausblenden', 'Elo-Spalte im Frontend ausblenden');
+$GLOBALS['TL_LANG']['tl_content']['kaderliste_stufen'] = array('Kaderstufen wählen', 'Wählen Sie ein oder mehrere Kaderstufen aus, die angezeigt werden sollen. Wenn Sie keine auswählen werden alle Stufen angezeigt.');